docs: Merge example app into single file so usage of flame_gamepads is in main.dart
